Remove duplicated `TreeNode` definition in physical-expr
