+++ datafusion/core/src/physical_plan/joins/hash_join_utils.rs
@@ -24,10 +24,10 @@ use std::usize;
 
 use arrow::datatypes::SchemaRef;
 
+use datafusion_common::tree_node::TreeNode;
 use datafusion_common::DataFusionError;
 use datafusion_physical_expr::expressions::Column;
 use datafusion_physical_expr::intervals::Interval;
-use datafusion_physical_expr::tree_node::TreeNode as PhysicalExprTreeNode;
 use datafusion_physical_expr::utils::collect_columns;
 use datafusion_physical_expr::{PhysicalExpr, PhysicalSortExpr};
