+++ datafusion/core/src/physical_plan/joins/sort_merge_join.rs
@@ -54,7 +54,7 @@ use crate::physical_plan::{
     Partitioning, PhysicalExpr, RecordBatchStream, SendableRecordBatchStream, Statistics,
 };
 
-use datafusion_physical_expr::tree_node::TreeNode as PhysicalExprTreeNode;
+use datafusion_common::tree_node::TreeNode;
 
 /// join execution plan executes partitions in parallel and combines them into a set of
 /// partitions.
